feature: add MXNet 1.4.1 Dockerfiles (#38)
starting with MXNet 1.4.1, the Python 3 CPU and GPU images will
be compatible with a variety of platforms: SageMaker, EKS, EC2.
The Python 2 and EI ones, however, will not. That is why
the Dockerfiles in this PR diverge so much between Python 2 and Python 3.
In addition, there is no MMS base image with CUDA 10.0,
so the Python 2 GPU image does change from the MXNet 1.4.0 version.
For now, I've left it in the release buildspec to keep
building and releasing Python 3 images as a sanity check,
but in the future we may elect to remove that.